requirements.txt
streamlit
fastapi
uvicorn
langchain
openai
PyPDF2
python-dotenv
